requirements.txt
# Core Dependencies
neo4j==5.16.0
llama-index==0.10.0
llama-index-graph-stores-neo4j==0.2.0
llama-index-embeddings-openai==0.1.6
llama-index-llms-openai==0.1.13

# LangChain/LangGraph Orchestration
langgraph==0.0.26
langchain==0.1.4
langchain-openai==0.0.5
langsmith==0.0.87

# Web Framework
streamlit==1.31.0
plotly==5.18.0
pandas==2.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
requests==2.31.0
